Created EnterSales Method.
